requirements.txt
scikit-learn
PyYAML
-e .